UHF-7029: UHF-7486: Move translations to separate files
