+++ src/js/calculator/daycare_payment/daycare_payment.js
@@ -1,4 +1,5 @@
-/* eslint-disable no-template-curly-in-string */
+import translations from './_translations';
+
 class DaycarePayment {
   constructor(id, settings) {
     this.id = id;
@@ -320,210 +321,6 @@ class DaycarePayment {
         },
       ]
     });
-
-    // Translations
-    const translations = {
-      family_info: {
-        fi: 'Perheen tiedot',
-        sv: null,
-        en: null,
-      },
-      household_size: {
-        fi: 'Perheen koko',
-        sv: null,
-        en: null,
-      },
-      household_size_is_too_small_for_child_count: {
-        fi: '${labelLink} pitäisi olla vähintään ${minValue} jos samassa osoitteessa asuu aikuinen ja ${childCount} lasta',
-        sv: null,
-        en: null,
-      },
-      household_size_explanation: {
-        fi: 'Samassa osoitteessa asuvien aikuisten ja alaikäisten lasten määrä.',
-        sv: null,
-        en: null,
-      },
-      gross_income_per_month: {
-        fi: 'Perheen tulot kuukaudessa',
-        sv: null,
-        en: null,
-      },
-      gross_income_per_month_explanation: {
-        fi: 'Koko perheen yhteenlasketut ansio- ja pääomatulot sekä etuudet ennen verojen vähentämistä.',
-        sv: null,
-        en: null,
-      },
-      child_info: {
-        fi: 'Lasten tiedot',
-        sv: null,
-        en: null,
-      },
-      child_info_paragraph: {
-        fi: 'Täytä tiedot koko perheen varhaiskasvatuksessa olevista tai sinne tulevista lapsista. Jos lapsia on useampi, maksu määräytyy nuorimmasta alkaen.',
-        sv: null,
-        en: null,
-      },
-      add_next_child: {
-        fi: 'Lisää seuraavan lapsen tiedot',
-        sv: null,
-        en: null,
-      },
-      remove_child: {
-        fi: 'Poista lapsen tiedot',
-        sv: null,
-        en: null,
-      },
-      youngest_child_title: {
-        fi: 'Nuorin tai ainoa varhaiskasvatuksessa oleva lapsi',
-        sv: null,
-        en: null,
-      },
-      nth_child_title: {
-        fi: 'Seuraavaksi nuorin varhaiskasvatuksessa oleva lapsi',
-        sv: null,
-        en: null,
-      },
-      daycare_type: {
-        fi: 'Varhaiskasvatuksen muoto',
-        sv: null,
-        en: null,
-      },
-      daycare_type_1: {
-        fi: 'Varhaiskasvatus arkipäivisin',
-        sv: null,
-        en: null,
-      },
-      daycare_type_2: {
-        fi: '6-vuotiaiden esiopetus ja varhaiskasvatus',
-        sv: null,
-        en: null,
-      },
-      daycare_type_3: {
-        fi: '5-vuotiaiden varhaiskasvatus',
-        sv: null,
-        en: null,
-      },
-      daycare_type_4: {
-        fi: 'Vuorohoito (hoitoa myös iltaisin ja viikonloppuisin)',
-        sv: null,
-        en: null,
-      },
-      daycare_type_1_caretime: {
-        fi: 'Hoitoaika',
-        sv: null,
-        en: null,
-      },
-      daycare_type_1_caretime_1: {
-        fi: 'Kokopäiväinen, yli 7 tuntia päivässä',
-        sv: null,
-        en: null,
-      },
-      daycare_type_1_caretime_2: {
-        fi: 'Kokopäiväinen, yli 5 ja enintään 7 tuntia päivässä',
-        sv: null,
-        en: null,
-      },
-      daycare_type_1_caretime_3: {
-        fi: 'Osapäiväinen, enintään 5 tuntia päivässä',
-        sv: null,
-        en: null,
-      },
-      daycare_free_days: {
-        fi: 'Vapaapäivien määrä',
-        sv: null,
-        en: null,
-      },
-      daycare_free_days_explanation: {
-        fi: 'Mahdolliset säännöllisesti toistuvat arkipäivät, jolloin lapsi ei osallistu päiväkotihoitoon. Kuukaudessa 4-12 vapaapäivää vaikuttaa maksuun.',
-        sv: null,
-        en: null,
-      },
-      daycare_type_2_caretime: {
-        fi: 'Esiopetuksen ja päiväkotihoidon hoitoaika yhteensä',
-        sv: null,
-        en: null,
-      },
-      daycare_type_2_caretime_1: {
-        fi: 'Yli 7 tuntia päivässä',
-        sv: null,
-        en: null,
-      },
-      daycare_type_2_caretime_2: {
-        fi: '7-8 tuntia päivässä (6 –vuotiaiden esiopetusikäisten ja 5 -vuotiaiden esiopetuskokeilu)',
-        sv: null,
-        en: null,
-      },
-      daycare_type_2_caretime_3: {
-        fi: 'Yli 5 ja alle 7 tuntia päivässä',
-        sv: null,
-        en: null,
-      },
-      daycare_type_2_caretime_4: {
-        fi: 'Enintään 5 tuntia päivässä',
-        sv: null,
-        en: null,
-      },
-      daycare_type_3_caretime: {
-        fi: 'Hoitoaika',
-        sv: null,
-        en: null,
-      },
-      daycare_type_3_caretime_1: {
-        fi: 'Yli 7 tuntia päivässä',
-        sv: null,
-        en: null,
-      },
-      daycare_type_3_caretime_2: {
-        fi: '5-7 tuntia päivässä',
-        sv: null,
-        en: null,
-      },
-      daycare_type_3_caretime_3: {
-        fi: '4-5 tuntia päivässä',
-        sv: null,
-        en: null,
-      },
-      daycare_type_3_caretime_4: {
-        fi: '4 tuntia päivässä (maksuton)',
-        sv: null,
-        en: null,
-      },
-      daycare_type_4_caretime: {
-        fi: 'Hoitoaika',
-        sv: null,
-        en: null,
-      },
-      daycare_type_4_caretime_1: {
-        fi: 'Yli 160 tuntia kuukaudessa',
-        sv: null,
-        en: null,
-      },
-      daycare_type_4_caretime_2: {
-        fi: '101-160 tuntia kuukaudessa',
-        sv: null,
-        en: null,
-      },
-      daycare_type_4_caretime_3: {
-        fi: '61-100 tuntia kuukaudessa',
-        sv: null,
-        en: null,
-      },
-      daycare_type_4_caretime_explanation: {
-        fi: 'Lapsella tulisi olla vähintään kaksi vapaapäivää viikossa tai vaihtoehtoisesti kahdeksan vapaapäivää kuukaudessa.',
-        sv: null,
-        en: null,
-      },
-      default_info_title: {
-        fi: 'Täytä kaikki pakolliset tiedot, ja paina "Laske arvio".',
-        sv: null,
-        en: null,
-      },
-      default_info_message: {
-        fi: 'Arvio asiakasmaksusta tulee näkyviin tähän. Jos päivität laskurin sisältöä, muista painaa "Laske arvio" uudelleen.',
-        sv: null,
-        en: null,
-      },
-    };
 
     const updateChild = (childNumber) => {
       const daycareTypeForChild = this.calculator.getFieldValue(`daycare_type_for_child_${childNumber}`);
