+++ src/js/calculator/home_care_client_payment/_translations.js
@@ -1,0 +1,230 @@
+/* eslint-disable no-template-curly-in-string */
+
+const translations = {
+  household_size: {
+    fi: 'Perheenjäsenten määrä numeroin',
+    sv: null,
+    en: null,
+  },
+  monthly_usage: {
+    fi: 'Kotona tarvittavan palvelun määrä tunteina kuukaudessa (palvelu - ja hoitosuunnitelman mukainen tuntimäärä)',
+    sv: null,
+    en: null,
+  },
+  monthly_usage_explanation: {
+    fi: 'Kuukausipalvelun määrä saadaan laskukaavalla: viikkotuntimäärä jaettuna seitsemällä x 30.',
+    sv: null,
+    en: null,
+  },
+  gross_income_per_month: {
+    fi: 'Bruttotulot kuukaudessa',
+    sv: null,
+    en: null,
+  },
+  gross_income_per_month_explanation: {
+    fi: 'jos ette halua ilmoittaa tulojanne, jättäkää kohta tyhjäksi.',
+    sv: null,
+    en: null,
+  },
+
+
+  phone: {
+    fi: 'Haluatteko käyttöönne Helsingin palvelukeskuksen turvapuhelimen?',
+    sv: null,
+    en: null,
+  },
+  phone_no: {
+    fi: 'Ei',
+    sv: null,
+    en: null,
+  },
+  phone_yes: {
+    fi: 'Kyllä',
+    sv: null,
+    en: null,
+  },
+  support_service_payments: {
+    fi: 'Tukipalvelujen maksut',
+    sv: null,
+    en: null,
+  },
+  support_service_payments_not_income_based: {
+    fi: 'Alla mainittujen tukipalvelujen maksut eivät ole tulosidonnaisia.',
+    sv: null,
+    en: null,
+  },
+  sauna_no_travel: {
+    fi: 'Saunotus ilman kuljetusta',
+    sv: null,
+    en: null,
+  },
+  sauna_with_travel: {
+    fi: 'Saunotus ja kuljetusmaksu',
+    sv: null,
+    en: null,
+  },
+  shopping: {
+    fi: 'Kauppapalvelu',
+    sv: null,
+    en: null,
+  },
+  foodservice: {
+    fi: 'Valitse ateriapalvelu',
+    sv: null,
+    en: null,
+  },
+  foodservice_1: {
+    fi: 'Olen eteläisen tai kaakkoisen kotihoidon asiakas',
+    sv: null,
+    en: null,
+  },
+  foodservice_2: {
+    fi: 'Olen muun kotihoitoyksikön asiakas',
+    sv: null,
+    en: null,
+  },
+  foodservice_3: {
+    fi: 'Käytössäni on Menumat-ateria-automaatti',
+    sv: null,
+    en: null,
+  },
+  foodservice_4: {
+    fi: 'Ei ateriapalvelua',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_1_delivery_fee: {
+    fi: 'Aterioiden kuljetusmaksu',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_1_delivery_fee_1: {
+    fi: 'Kylmien aterioiden kuljetus 4 krt / kk, á 7,21',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_1_delivery_fee_2: {
+    fi: 'Kylmien aterioiden kuljetus 8 krt / kk, á 7,21',
+    sv: null,
+    en: null,
+  },
+  meal_price: {
+    fi: 'Aterian hinta',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_1_meal_price_1: {
+    fi: '4,98 €',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_1_meal_price_2: {
+    fi: '5,49 €',
+    sv: null,
+    en: null,
+  },
+  meals: {
+    fi: 'Ateriat',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_2_delivery_fee: {
+    fi: 'Aterioiden kuljetusmaksu',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_2_delivery_fee_1: {
+    fi: 'Kylmien aterioiden kuljetus 4 krt / kk, á 7,21',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_2_delivery_fee_2: {
+    fi: 'Kylmien aterioiden kuljetus 8 krt / kk, á 7,21',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_2_meals: {
+    fi: 'Ateriat (määrä / kk)',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_3_meal_price: {
+    fi: 'Aterian hinta',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_3_meal_price_1: {
+    fi: 'Lounas 4,99 €',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_3_meal_price_2: {
+    fi: 'Lounas 4,99 € + jälkiruoka 1,12 €',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_3_meal_price_3: {
+    fi: 'Lounas 5,57 €',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_3_meal_price_4: {
+    fi: 'Lounas 5,57 € + jälkiruoka 1,12 €',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_3_usage: {
+    fi: 'Palvelun käyttö',
+    sv: null,
+    en: null,
+  },
+  foodservice_group_3_meals: {
+    fi: 'Ateriat',
+    sv: null,
+    en: null,
+  },
+
+
+  info_voucher_value: {
+    fi: 'Palvelusetelin arvo ${voucher_value} € / kk.',
+    sv: null,
+    en: null,
+  },
+  alert_voucher_not_applicable: {
+    fi: 'Ei oikeutta palveluseteliin.',
+    sv: null,
+    en: null,
+  },
+  error_check_size: {
+    fi: 'Tarkista "Talouden koko"',
+    sv: null,
+    en: null,
+  },
+  error_check_usage: {
+    fi: 'Tarkista "Kuukausittainen käyttö"',
+    sv: null,
+    en: null,
+  },
+  error_check_income: {
+    fi: 'Tarkista "Bruttotulot kuukaudessa"',
+    sv: null,
+    en: null,
+  },
+  error_check_price: {
+    fi: 'Tarkista "Palveluntuottajan tuntihinta"',
+    sv: null,
+    en: null,
+  },
+  missing_input: {
+    fi: 'Tietoja puuttuu',
+    sv: null,
+    en: null,
+  },
+  result: {
+    fi: 'Lopputulos',
+    sv: null,
+    en: null,
+  },
+};
+
+export default translations;
