+++ src/js/calculator/home_care_service_voucher/home_care_service_voucher.js
@@ -1,4 +1,5 @@
-/* eslint-disable no-template-curly-in-string */
+import translations from './_translations';
+
 class HomeCareServiceVoucher {
   constructor(id, settings) {
     this.id = id;
@@ -60,60 +61,6 @@ class HomeCareServiceVoucher {
           },
         ]
       });
-
-    // Translations
-    const translations = {
-      household_size: {
-        fi: 'Talouden koko',
-        sv: null,
-        en: null,
-      },
-      household_size_explanation: {
-        fi: 'Taloudella tarkoitetaan samassa osoitteessa yhteistaloudessa eläviä aikuisia sekä heidän alaikäisiä lapsiaan',
-        sv: null,
-        en: null,
-      },
-      monthly_usage: {
-        fi: 'Kuukausittainen käyttö',
-        sv: null,
-        en: null,
-      },
-      monthly_usage_explanation: {
-        fi: 'Palvelusuunnitelmassa teille määritelty kotihoidon tuntimäärä kuukaudessa.',
-        sv: null,
-        en: null,
-      },
-      gross_income_per_month: {
-        fi: 'Bruttotulot kuukaudessa',
-        sv: null,
-        en: null,
-      },
-      gross_income_per_month_explanation: {
-        fi: 'Kuukausituloina huomioidaan veronalaiset ansiotulot ja veronalaiset pääomatulot, joita ovat metsätulot sekä korko-, osinko- ja vuokratulot.',
-        sv: null,
-        en: null,
-      },
-      service_providers_hourly_price: {
-        fi: 'Palveluntuottajan tuntihinta',
-        sv: null,
-        en: null,
-      },
-      service_providers_hourly_price_explanation: {
-        fi: 'Valitsemanne palveluntuottajan perimä kotihoitotunnin hinta',
-        sv: null,
-        en: null,
-      },
-      info_voucher_value: {
-        fi: 'Palvelusetelin arvo ${voucher_value} € / kk.',
-        sv: null,
-        en: null,
-      },
-      alert_voucher_not_applicable: {
-        fi: 'Ei oikeutta palveluseteliin.',
-        sv: null,
-        en: null,
-      },
-    };
 
     const validate = () => {
       const errorMessages = [];
