+++ src/js/calculator/house_cleaning_service_voucher/_translations.js
@@ -1,0 +1,46 @@
+/* eslint-disable no-template-curly-in-string */
+
+const translations = {
+  gross_income_per_month: {
+    fi: 'Bruttotulot kuukaudessa',
+    sv: 'Bruttoinkomst per månad',
+    en: 'Gross income per month',
+  },
+  household_size: {
+    fi: 'Talouden koko',
+    sv: 'Storleken på hushållet',
+    en: 'Household size',
+  },
+  one_person_household: {
+    fi: 'Yhden hengen talous',
+    sv: 'Enmanshushåll',
+    en: 'One person household',
+  },
+  two_person_household: {
+    fi: 'Kahden hengen talous',
+    sv: 'Två personers hushåll',
+    en: 'Two person household',
+  },
+  info_voucher_value: {
+    fi: 'Palvelusetelin arvo ${voucher_value} € / kk.',
+    sv: 'Värde på servicekupong ${voucher_value} € / månad',
+    en: 'Service voucher value ${voucher_value} € / month',
+  },
+  alert_voucher_not_applicable: {
+    fi: 'Ei oikeutta palveluseteliin',
+    sv: 'Ingen rätt till servicekupong',
+    en: 'No right to a service voucher',
+  },
+  error_gross_income_per_month: {
+    fi: 'Tarkista "Bruttotulot kuukaudessa"',
+    sv: 'Kolla upp "Bruttoinkomst per månad"',
+    en: 'Please check "Gross income per month"',
+  },
+  error_household_size: {
+    fi: 'Tarkista "Talouden koko"',
+    sv: 'Kolla upp "Storleken på hushållet"',
+    en: 'Please check "Household size"',
+  },
+};
+
+export default translations;
